Add audio.cpp
